rename insecure_mcp to local_http_stdio
