+++ docs/release_notes/v1.1.4.md
@@ -1,0 +1,28 @@
+## [1.1.4] - 2026-06-19
+
+### Overview
+
+v1.1.4 is a code quality and test coverage release that significantly improves the reliability and maintainability of the g8e platform. This release focuses on comprehensive MCP tool test refactoring, integration test improvements, database error handling cleanup, and cross-platform path handling enhancements.
+
+### Added
+
+* **Comprehensive MCP Tool Tests** — Added extensive unit tests for file read operations, disk usage monitoring (including Windows-specific implementations), and network socket auditing.
+* **Network Socket Audit Tool** — Added new `net_socket_audit` MCP tool for inspecting active network connections and socket states.
+* **Field Parser Enhancements** — Enhanced field parsing capabilities with improved validation and error handling.
+
+### Changed
+
+* **MCP Tool Refactoring** — Systematic refactoring of MCP native tools for improved testability and dependency injection support.
+* **Integration Test Improvements** — Refactored integration tests for better separation of concerns and improved reliability.
+* **Database Error Handling** — Cleaned up and standardized database error handling across storage services.
+* **Keystore Service Improvements** — Enhanced keystore backend implementations with better error handling and cross-platform support.
+* **Gateway State Root Query** — Fixed and improved gateway state root query functionality with associated test improvements.
+* **Demo SOC Refactoring** — Refactored demo configurations for cleaner separation of concerns (SOC).
+* **Bootstrap Function Split** — Improved bootstrap function organization by splitting responsibilities.
+
+### Fixed
+
+* **Windows Path Handling** — Improved Windows path handling across multiple services for better cross-platform compatibility.
+* **Demo Error Handling** — Fixed error handling in demo configurations.
+* **Code Quality** — Applied gofmt and addressed various code quality issues identified by linters.
+* **Pubsub Commands** — Improved pubsub command handling and results processing.
