+++ internal/constants/ports_test.go
@@ -28,14 +28,14 @@ func TestPorts(t *testing.T) {
 		assert.Equal(t, 8443, Ports.OperatorHttps)
 	})
 
-	t.Run("insecure mcp gateway port has correct value", func(t *testing.T) {
-		assert.Equal(t, 18789, Ports.InsecureMcpGateway)
+	t.Run("local http stdio gateway port has correct value", func(t *testing.T) {
+		assert.Equal(t, 18789, Ports.LocalHttpStdioGateway)
 	})
 
 	t.Run("snake case variants match camel case", func(t *testing.T) {
 		assert.Equal(t, Ports.OperatorHttp, Ports.OperatorHttpSnake)
 		assert.Equal(t, Ports.OperatorHttps, Ports.OperatorHttpsSnake)
-		assert.Equal(t, Ports.InsecureMcpGateway, Ports.InsecureMcpGatewaySnake)
+		assert.Equal(t, Ports.LocalHttpStdioGateway, Ports.LocalHttpStdioGatewaySnake)
 	})
 
 	t.Run("all ports are in valid range", func(t *testing.T) {
@@ -43,13 +43,13 @@ func TestPorts(t *testing.T) {
 		assert.LessOrEqual(t, Ports.OperatorHttp, 65535)
 		assert.GreaterOrEqual(t, Ports.OperatorHttps, 1)
 		assert.LessOrEqual(t, Ports.OperatorHttps, 65535)
-		assert.GreaterOrEqual(t, Ports.InsecureMcpGateway, 1)
-		assert.LessOrEqual(t, Ports.InsecureMcpGateway, 65535)
+		assert.GreaterOrEqual(t, Ports.LocalHttpStdioGateway, 1)
+		assert.LessOrEqual(t, Ports.LocalHttpStdioGateway, 65535)
 	})
 
 	t.Run("ports are distinct", func(t *testing.T) {
 		assert.NotEqual(t, Ports.OperatorHttp, Ports.OperatorHttps)
-		assert.NotEqual(t, Ports.OperatorHttp, Ports.InsecureMcpGateway)
-		assert.NotEqual(t, Ports.OperatorHttps, Ports.InsecureMcpGateway)
+		assert.NotEqual(t, Ports.OperatorHttp, Ports.LocalHttpStdioGateway)
+		assert.NotEqual(t, Ports.OperatorHttps, Ports.LocalHttpStdioGateway)
 	})
 }
