+++ internal/services/local_http_stdio/insecure_mcp_node_service.go
@@ -11,7 +11,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package insecure_mcp
+package local_http_stdio
 
 import (
 	"bytes"
