+++ protocol/constants/ports.json
@@ -1,9 +1,9 @@
 {
   "ports": {
-    "InsecureMcpGateway": 18789,
+    "LocalHttpStdioGateway": 18789,
     "OperatorHttp": 8080,
     "OperatorHttps": 8443,
-    "insecure_mcp_gateway": 18789,
+    "local_http_stdio_gateway": 18789,
     "operator_http": 8080,
     "operator_https": 8443
   }
