+++ protocol/python/pyproject.toml
@@ -17,7 +17,7 @@ build-backend = "setuptools.build_meta"
 
 [project]
 name = "g8e-protocol"
-version = "1.1.3"
+version = "1.1.4"
 description = "g8e Protocol Constants and Models"
 readme = "README.md"
 requires-python = ">=3.10"
